refactor: align to Strapi Plugin SDK
