+++ .npmignore
@@ -4,7 +4,7 @@ coverage/
 .github/
 .nvmrc
 .eslintrc
-dist/tsconfig.tsbuildinfo
+.eslintignore
 codecov.yml
 *.map
 *.spec.*
