+++ admin/custom.d.ts
@@ -1,4 +1,2 @@
 declare module '@strapi/design-system/*';
 declare module '@strapi/design-system';
-declare module '@strapi/icons';
-declare module '@strapi/icons/*';
