+++ admin/src/pluginId.ts
@@ -1,0 +1,1 @@
+export const pluginId = "imgix";
