--- pluginRepository.ts
@@ -1,3 +1,0 @@
-import * as pluginPkg from './package.json';
-
-export const pluginRepository = pluginPkg.repository.url;
