+++ server/src/register.ts
