+++ server/src/utils/getService.ts
@@ -1,6 +1,6 @@
 import { Core } from '@strapi/types';
 
-import { pluginId } from '../utils';
+import { pluginId } from '.';
 import { PluginServices } from '../services';
 
 export const getService = <ServiceName extends keyof PluginServices>(strapi: Core.Strapi, serviceName: ServiceName): PluginServices[ServiceName] => {
