+++ server/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './getService';
+export * from '../../../admin/src/pluginId';
