--- server/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './getService';
-export * from './pluginId';
