--- server/utils/pluginId.ts
@@ -1,3 +1,0 @@
-import * as pluginPkg from './../../package.json';
-
-export const pluginId = pluginPkg.name.replace(/^strapi-plugin-/i, "");
