--- strapi-admin.ts
@@ -1,5 +1,0 @@
-'use strict';
-
-import admin from './admin/src';
-
-export default admin;
