--- strapi-server.ts
@@ -1,5 +1,0 @@
-'use strict';
-
-import server from './server';
-
-export default server;
