--- tsconfig.admin.json
@@ -1,41 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./",
-    "outDir": "dist",
-    "target": "esnext",
-    "module": "esnext",
-    "jsx": "react",
-    "moduleResolution": "bundler",
-    "useDefineForClassFields": true,
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "removeComments": true,
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "downlevelIteration": true,
-    "incremental": true,
-    "allowJs": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true
-  },
-
-  "include": [
-    "custom.d.ts",
-    "admin/**/*",
-    "strapi-admin.*",
-    "setup-package.ts",
-  ],
-  "exclude": [
-    "**/node_modules/**/*",
-    "**/__tests__/**/*",
-    "**/__mocks__/**/*",
-    "**/dist/**/*",
-    "strapi-server.*",
-    "server/**/*",
-  ]
-}
