--- tsconfig.server.json
@@ -1,39 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./",
-    "outDir": "dist",
-    "target": "esnext",
-    "module": "commonjs",
-    "jsx": "react",
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "removeComments": true,
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "downlevelIteration": true,
-    "incremental": true,
-    "allowJs": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true
-  },
-
-  "include": [
-    "server/**/*",
-    "strapi-server.*",
-  ],
-  "exclude": [
-    "**/node_modules/**/*",
-    "**/__tests__/**/*",
-    "**/__mocks__/**/*",
-    "**/dist/**/*",
-    "custom.d.ts",
-    "admin/**/*",
-    "strapi-admin.*",
-    "setup-package.ts",
-  ]
-}
